# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FSTS Client Dashboard Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all dashboard users and their site role assignments
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DashboardUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Invite / create a dashboard user with a role on one or more sites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardUser"
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user's role assignments or status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardUser"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Remove a dashboard user (requires confirmation client-side)
      responses:
        "204":
          description: Deleted
...
  /sites/{siteId}/square-config:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getSquareConfig
      tags: [square]
      summary: Get Square payment configuration for a site
      responses:
        "200":
          description: Square config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SquareConfig"
    put:
      operationId: updateSquareConfig
      tags: [square]
      summary: Update Square payment configuration for a site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SquareConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SquareConfig"
...
components:
  schemas:
...
    Role:
      type: string
      enum:
        - super_admin
        - client_admin
        - editor
        - marketing
        - training_manager
        - read_only
    SiteRoleAssignment:
      type: object
      required: [siteId, siteName, role]
      properties:
        siteId: { type: integer }
        siteName: { type: string }
        role:
          $ref: "#/components/schemas/Role"
...
    DashboardUser:
      type: object
      required: [id, clerkUserId, name, email, isSuperAdmin, isActive, roleAssignments, createdAt]
      properties:
        id: { type: integer }
        clerkUserId: { type: string }
        name: { type: string }
        email: { type: string }
        isSuperAdmin: { type: boolean }
        isActive: { type: boolean }
        roleAssignments:
          type: array
          items:
            $ref: "#/components/schemas/SiteRoleAssignment"
        createdAt: { type: string, format: date-time }
    UserRoleAssignmentInput:
      type: object
      required: [siteId, role]
      properties:
        siteId: { type: integer }
        role:
          $ref: "#/components/schemas/Role"
    UserInput:
      type: object
      required: [name, email]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        isSuperAdmin: { type: boolean }
        roleAssignments:
          type: array
          items:
            $ref: "#/components/schemas/UserRoleAssignmentInput"
    UserUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        isSuperAdmin: { type: boolean }
        isActive: { type: boolean }
        roleAssignments:
          type: array
          items:
            $ref: "#/components/schemas/UserRoleAssignmentInput"
...
    SquareConfig:
      type: object
      required: [siteId, connected, environment, updatedAt]
      properties:
        siteId: { type: integer }
        connected: { type: boolean }
        environment:
          type: string
          enum: [sandbox, production]
        applicationIdLast4: { type: ["string", "null"] }
        locationId: { type: ["string", "null"] }
        checkoutEnabled: { type: boolean }
        updatedAt: { type: string, format: date-time }
    SquareConfigInput:
      type: object
      required: [environment]
      properties:
        environment:
          type: string
          enum: [sandbox, production]
        applicationId: { type: string }
        locationId: { type: string }
        accessToken: { type: string }
        checkoutEnabled: { type: boolean }
...