# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FSTS Client Dashboard Platform API
 * OpenAPI spec version: 0.1.0
 */

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  super_admin: 'super_admin',
  client_admin: 'client_admin',
  editor: 'editor',
  marketing: 'marketing',
  training_manager: 'training_manager',
  read_only: 'read_only',
} as const;